Extract EventsAwaitingQueries
 - extracting EventsBuilder as standalone class
 - extracting TestingEventListener as standalone class
