+++ presto-tests/src/test/java/io/prestosql/execution/TestEventListenerPlugin.java
@@ -14,13 +14,9 @@
 package io.prestosql.execution;
 
 import com.google.common.collect.ImmutableList;
-import io.prestosql.execution.TestEventListener.EventsBuilder;
 import io.prestosql.spi.Plugin;
 import io.prestosql.spi.eventlistener.EventListener;
 import io.prestosql.spi.eventlistener.EventListenerFactory;
-import io.prestosql.spi.eventlistener.QueryCompletedEvent;
-import io.prestosql.spi.eventlistener.QueryCreatedEvent;
-import io.prestosql.spi.eventlistener.SplitCompletedEvent;
 
 import java.util.Map;
 
@@ -65,35 +61,6 @@ public class TestEventListenerPlugin
         public EventListener create(Map<String, String> config)
         {
             return new TestingEventListener(eventsBuilder);
-        }
-    }
-
-    private static class TestingEventListener
-            implements EventListener
-    {
-        private final EventsBuilder eventsBuilder;
-
-        public TestingEventListener(EventsBuilder eventsBuilder)
-        {
-            this.eventsBuilder = eventsBuilder;
-        }
-
-        @Override
-        public void queryCreated(QueryCreatedEvent queryCreatedEvent)
-        {
-            eventsBuilder.addQueryCreated(queryCreatedEvent);
-        }
-
-        @Override
-        public void queryCompleted(QueryCompletedEvent queryCompletedEvent)
-        {
-            eventsBuilder.addQueryCompleted(queryCompletedEvent);
-        }
-
-        @Override
-        public void splitCompleted(SplitCompletedEvent splitCompletedEvent)
-        {
-            eventsBuilder.addSplitCompleted(splitCompletedEvent);
         }
     }
 }
